+++ js/main.js
@@ -1,4 +1,3 @@
-import { findPicture, pictures } from '/js/popup.js';
+import { findPicture, pictures } from '/js/full-picture.js';
 
 findPicture(pictures);
-
